check the queue and qsub the file
Due to the teacher myw has limited the number of jobs that we can submit at the new students' services, I wrote the file to check the queue every two minutes by the command 'qstat'. And if the queue is not full, the program will qsub the file. The check will do over and over again until all the jobs has been submited. But because I don't know how to use nohup in the shell script, so the program must be use the format as "nohup queuesub [your file] &".